Include Bindings Panel with hooks instead
