+++ packages/block-editor/src/hooks/block-bindings-panel.scss
